Fixed potential infinite recursion and ultrasonic
fixed the broken sensor cable for the ultrasonic sensor and added code to convert from mm to in.
Also fixed potential infitite recursion in gocube.
gocube() calls turncube() which calls gocube() ... Im sure you can see where this is going. Anyways I fixed it be editing the conditions for the while loop