Fixed Ultrasonic Readings, and added GoPID loop
I ultrasonic was returning wack readings since I was multiplying by 2.54, rather than dividing when converting from mm to in. A week of debugging because of 1(!) stupid character. After that I copypasted the PID loop from turncube in to goCube I still need to tune it some as its v. Jerky at the moment.