Tuned PIDs, added more auton, updated .gitignore
I finally got around the tuning the PID and made Vision significantly better. Im having some trouble with going to the ultrasonic cube since if it doesnt see it, it will continue without stopping which is an issue. I added some more auton to test it since this is what ive been working on. Finally I change .gitignore to get rid of some useless files.