Auton is broken adding new cube sensing mechanism
Our PID loop although it works is very unreliable since if it slips and the sonar sensor detects something other than the cube it will go ape and never stop. So im trying to add a sensor beneath the claw to sense when a cube is in it. also made some classes child functions so welcome to this-> galore